requirements.txt
django==5.0.6
nltk==3.8.1
python-decouple==3.8
djangorestframework==3.15.2
numpy==2.0.0